RAGs/Resume-Ranking-copilot/requirements.txt
streamlit
google-generativeai
llama-index
